Fix improper symbol visibility & weakness
We were forgetting to apply explicit weak/non-weak symbol lists that
the upsteam provides. This fixes certain symbols being weak when they
should not be. This matters if an executable ships its own copy of libc++,
whose symbols would conflict with the systems' copy (which is always
loaded into the process because of libSystem transitively depending on
it).

See https://github.com/darlinghq/darling/issues/879